# Speed up functions deployment