--- a/package.json
+++ b/package.json
@@ -44,7 +44,7 @@
   "homepage": "https://github.com/firebase/firebase-tools",
   "dependencies": {
     "JSONStream": "^1.2.1",
-    "archiver": "^0.16.0",
+    "archiver": "^2.1.0",
     "chalk": "^1.1.0",
     "cjson": "^0.3.1",
     "cli-table": "^0.3.1",