### Description

I've noticed that deploying firebase functions takes very long. The reason for this is that the package firebase-tools is using to archive the code into a zip archive is outdated. By upgrading the ```archiver``` package from ```^0.16.0``` to the newest version of ```^2.1.0``` the archive time can be drastically reduced.

Some stats:

Before:
```
i  functions: preparing ./ directory for uploading...
Start copying source
Copy source: 10521.333ms
Start packaging source
Reader listed 34638 files
Package source: 683353.692ms
i  functions: packaged ./ (177.87 MB) for uploading
```

After:
```
i  functions: preparing ./ directory for uploading...
Start copying source
Copy source: 9761.668ms
Start packaging source
Reader listed 34638 files
Package source: 41077.057ms
i  functions: packaged ./ (177.87 MB) for uploading
```

Archive time has been reduced from ```11.4 min``` to ```41 sec```